canary/http/requirements.txt
requests
kafka-python
python-dotenv
